fix: hexception optimize
