+++ client/src/ui/components/construction/GroundGrid.tsx
@@ -2,70 +2,31 @@ import useUIStore from "../../../hooks/store/useUIStore";
 import * as THREE from "three";
 import { createHexagonShape } from "../worldmap/hexagon/HexagonGeometry";
 import { HEX_RADIUS } from "../worldmap/hexagon/WorldHexagon";
-import { getEntityIdFromKeys, getUIPositionFromColRow } from "../../utils/utils";
-import { Bvh, useGLTF } from "@react-three/drei";
-import { useState } from "react";
+import { getUIPositionFromColRow } from "../../utils/utils";
+import { useMemo } from "react";
 import { useBuildingSound } from "../../../hooks/useUISound";
 import { useDojo } from "@/hooks/context/DojoContext";
 import useRealmStore from "@/hooks/store/useRealmStore";
-import { BuildingType, ResourcesIds, BuildingStringToEnum } from "@bibliothecadao/eternum";
-import { useQuery } from "@/hooks/helpers/useQuery";
-import { BuildingTooltip } from "../cityview/BuildingTooltip";
-import { Component, Has, HasValue, getComponentValue } from "@dojoengine/recs";
+import { BuildingType } from "@bibliothecadao/eternum";
 import { CairoOption, CairoOptionVariant } from "starknet";
-import { useEntityQuery } from "@dojoengine/react";
-
-// 1. Working and fairly smooth minimal lag and good transitions. And the construction OS box is nice
-// 2. Graphics are good and the hexagons are well placed
 
 export const isHexOccupied = (col: number, row: number, buildings: any[]) => {
   return buildings.some((building) => building.col === col && building.row === row);
 };
 
 const GroundGrid = () => {
-  const { hexPosition: globalHex } = useQuery();
+  const hexPositions = useMemo(() => generateHexPositions(), []);
 
   const { playBuildingSound } = useBuildingSound();
-  const hexPositions = generateHexPositions();
-  const {
-    previewBuilding,
-    hoveredBuildHex,
-    setHoveredBuildHex,
-    existingBuildings,
-    setExistingBuildings,
-    selectedResource,
-    setResourceId,
-  } = useUIStore((state) => ({
-    previewBuilding: state.previewBuilding,
-    hoveredBuildHex: state.hoveredBuildHex,
-    setHoveredBuildHex: state.setHoveredBuildHex,
-    existingBuildings: state.existingBuildings,
-    setExistingBuildings: state.setExistingBuildings,
-    selectedResource: state.selectedResource,
-    setResourceId: state.setResourceId,
-  }));
+  const { previewBuilding, setHoveredBuildHex, existingBuildings, selectedResource } = useUIStore((state) => state);
   const { realmEntityId } = useRealmStore();
 
   const {
     account: { account },
     setup: {
       systemCalls: { create_building },
-      components: { Building },
     },
   } = useDojo();
-
-  const models = useGLTF([
-    "/models/buildings/castle.glb",
-    "/models/buildings/farm.glb",
-    // "/models/buildings/fishery.glb",
-    "/models/buildings/mine.glb",
-    "/models/buildings/stable.glb",
-    "/models/buildings/workhut.glb",
-    "/models/buildings/archer_range.glb",
-    "/models/buildings/barracks.glb",
-    "/models/buildings/market.glb",
-    "/models/buildings/storehouse.glb",
-  ]);
 
   const handlePlacement = async (col: number, row: number, previewBuilding: BuildingType) => {
     await create_building({
@@ -83,101 +44,46 @@ const GroundGrid = () => {
     });
   };
 
-  const builtBuildings = useEntityQuery([
-    Has(Building),
-    HasValue(Building, { outer_col: BigInt(globalHex.col), outer_row: BigInt(globalHex.row) }),
-  ]);
-
   return (
     <>
-      {builtBuildings.map((entity) => {
-        const productionModelValue = getComponentValue(Building, entity);
-
-        return (
-          productionModelValue && (
-            <BuiltBuilding
-              key={entity}
-              models={models}
-              buildingCategory={
-                BuildingStringToEnum[productionModelValue.category as keyof typeof BuildingStringToEnum]
-              }
-              position={{ col: Number(productionModelValue.inner_col), row: Number(productionModelValue.inner_row) }}
-              onPointerMove={() =>
-                previewBuilding &&
-                setHoveredBuildHex({
-                  col: Number(productionModelValue.inner_col),
-                  row: Number(productionModelValue.inner_row),
-                })
-              }
-            />
-          )
-        );
-      })}
       <group rotation={[Math.PI / -2, 0, 0]} position={[0, 2, 0]}>
-        <Bvh firstHitOnly>
-          {hexPositions.map((hexPosition, index) => (
-            <Hexagon
-              key={index}
-              position={hexPosition}
-              onPointerMove={() =>
-                previewBuilding && setHoveredBuildHex({ col: hexPosition.col, row: hexPosition.row })
+        {hexPositions.map((hexPosition, index) => (
+          <Hexagon
+            key={index}
+            position={hexPosition}
+            onPointerEnter={() => previewBuilding && setHoveredBuildHex({ col: hexPosition.col, row: hexPosition.row })}
+            onClick={() => {
+              if (previewBuilding && !isHexOccupied(hexPosition.col, hexPosition.row, existingBuildings)) {
+                handlePlacement(hexPosition.col, hexPosition.row, previewBuilding);
+                playBuildingSound(previewBuilding);
               }
-              onClick={() => {
-                if (previewBuilding && !isHexOccupied(hexPosition.col, hexPosition.row, existingBuildings)) {
-                  handlePlacement(hexPosition.col, hexPosition.row, previewBuilding);
-                  playBuildingSound(previewBuilding);
-                }
-              }}
-            />
-          ))}
-        </Bvh>
+            }}
+          />
+        ))}
       </group>
     </>
   );
 };
 
-export const BuiltBuilding = ({
-  position,
-  onPointerMove,
-  models,
-  buildingCategory,
-}: {
-  position: any;
-  onPointerMove: any;
-  models: any;
-  buildingCategory: number;
-}) => {
-  const { x, y } = getUIPositionFromColRow(position.col, position.row, true);
-
-  return <primitive scale={3} object={models[buildingCategory].scene.clone()} position={[x, 2.33, -y]} />;
-};
-
 const hexagonGeometry = new THREE.ShapeGeometry(createHexagonShape(HEX_RADIUS));
 const mainColor = new THREE.Color(0.21389107406139374, 0.14227265119552612, 0.06926480680704117);
 const secondaryColor = mainColor.clone().lerp(new THREE.Color(1, 1, 1), 0.2);
+const mainMaterial = new THREE.MeshMatcapMaterial({ color: mainColor });
+const secondaryMaterial = new THREE.MeshMatcapMaterial({ color: secondaryColor });
 
-export const Hexagon = ({ position, onClick, onPointerMove }: { position: any; onClick: any; onPointerMove: any }) => {
-  const [showTooltip, setShowTooltip] = useState(false);
+export const Hexagon = ({
+  position,
+  onClick,
+  onPointerEnter,
+}: {
+  position: any;
+  onClick: any;
+  onPointerEnter: any;
+}) => {
   return (
-    <group position={[position.x, position.y, position.z]} onPointerMove={onPointerMove} onClick={onClick}>
-      <mesh geometry={hexagonGeometry} scale={0.5} position={[0, 0, 0.01]}>
-        <meshMatcapMaterial color={secondaryColor} />
-      </mesh>
-      {showTooltip ? <BuildingTooltip resourceId={10} /> : null}
-
-      <mesh
-        geometry={hexagonGeometry}
-        // onPointerEnter={(e) => {
-        //   e.stopPropagation();
-        //   setShowTooltip(true);
-        // }}
-        // onPointerLeave={(e) => {
-        //   e.stopPropagation();
-        //   setShowTooltip(false);
-        // }}
-      >
-        <meshMatcapMaterial color={mainColor} />
-      </mesh>
+    <group position={[position.x, position.y, position.z]} onPointerEnter={onPointerEnter} onClick={onClick}>
+      <mesh geometry={hexagonGeometry} scale={0.5} position={[0, 0, 0.01]} material={secondaryMaterial} />
+      <mesh geometry={hexagonGeometry} material={mainMaterial} />
     </group>
   );
 };
